projects/agents/InsightBot/requirements.txt
# Requirements for Analytical Chatbot project
python-dotenv==1.0.1
streamlit==1.28.0
langchain==0.2.14
langchain_community==0.2.12
langchain_cohere==0.3.0
tavily-python==0.5.0
transformers==4.44.2
torch==2.4.0
langchain_experimental==0.0.64
matplotlib==3.9.2
pytest==8.3.2